fixs all register controller bugs

+++ src/utils/cloudinary.js
@@ -16,15 +16,22 @@ import fs from "fs";
         // upload the file on cloudinary
        const response = await cloudinary.uploader.upload(localFilePath, {resource_type: "auto"})
         // file has been uploaded successfully
-        console.log("File is uploaded on cloudinary", response.url);
+        // console.log("File is uploaded on cloudinary", response.url);
+        
+        fs.unlinkSync(localFilePath)
         return response;
 
 
-      } catch (error) {
-        fs.unlinkSync(localFilePath); // remove the local save temporary file as the upload operation failed
-        return null;
-      }
+      }catch (error) {
+       console.error("Cloudinary Error:", error);
+
+      if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+
+    return null;
+      }
+  }
 
    export {uploadOnCloudinary}
